% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{action}
\alias{action}
\title{Specify an action in a stage}
\usage{
action(name, set, strategyMethodDomain = NULL, ...)
}
\arguments{
\item{name}{The variable name of the action}

\item{set}{The set of different action values. Can be a formula that depends on other game variables.}

\item{strategyMethodDomain}{if not NULL the action shall be specified via strategy method in an experiment. State the variable name upon which the action conditions
Only used when running an experiment or analysing experimental data}
}
\description{
Specify an action in a stage
}
\seealso{
Other Define Stage: \code{\link{natureMove}},
  \code{\link{stage}}
}
