% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{as.names.list}
\alias{as.names.list}
\title{Names lists are used to recursively store order of columns}
\usage{
as.names.list(names)
}
\description{
Names lists are used to recursively store order of columns
}
