% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_distinctions.r
\name{case_distinction}
\alias{case_distinction}
\title{A simple function to define case distinctions}
\usage{
case_distinction(...)
}
\description{
A simple function to define case distinctions
}
\seealso{
Other Helper Functions: \code{\link{is_true}}
}
