% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambit.r
\name{cond.eq.outcome}
\alias{cond.eq.outcome}
\title{return a conditional equilibrium outcome
cond is a list with variable names and their assumed value
we only pick rows from oco.df in which the condition is satisfied
we set the probabilities of the conditioned variable values to 1}
\usage{
cond.eq.outcome(eq.mat, cond, tg = NULL, oco.df = tg$oco.df,
  eq.ind = first.non.null(attr(eq.mat, "eq.ind"), NA),
  eo.df = eq.outcome(eq.mat = eq.mat, oco.df = oco.df, tg = tg),
  ceqo.ind = 1)
}
\description{
return a conditional equilibrium outcome
cond is a list with variable names and their assumed value
we only pick rows from oco.df in which the condition is satisfied
we set the probabilities of the conditioned variable values to 1
}
