% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{contains}
\alias{contains}
\title{Does an environment / list contain the objects named as names}
\usage{
contains(env, names, inherits = FALSE, ...)
}
\description{
Does an environment / list contain the objects named as names
}
