% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{custom.whisker.render}
\alias{custom.whisker.render}
\title{Adapts whisker render for different whisker formats}
\usage{
custom.whisker.render(template, data, ..., whiskers = c("<<", ">>"))
}
\description{
Adapts whisker render for different whisker formats
}
