% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{cut.to.sub.tree.path}
\alias{cut.to.sub.tree.path}
\title{Cuts away early stuff from a tree path}
\usage{
\method{cut}{to.sub.tree.path}(tree.path, after)
}
\description{
Cuts away early stuff from a tree path
}
