% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{eq_cond_outcomes}
\alias{eq_cond_outcomes}
\title{Return conditional equilibrium outcomes}
\usage{
eq_cond_outcomes(game, ..., fixed.list = list(...), fixed.vars = NULL)
}
\arguments{
\item{game}{the game object for which equilibria were computed e.g. with \code{game_solve}.}

\item{...}{variable names and their assumed value. We set the probabilities of the conditioned variable values to 1. These correspond to equilibrium outcomes given an unexpected tremble that makes the variables take the specified values. Variables can take multiple values. We then compute conditional equilibrium outcomes for each combination of values}

\item{fixed.list}{Alternativly to ... a named list with values to fix.}

\item{fixed.vars}{Alternative to ... or fixed.list, a vector of variable names. If provided, we compute the conditional expected outcomes holding fixed every possible combination of the variables stated in fixed.vars}
}
\description{
Return conditional equilibrium outcomes
}
\seealso{
Other eq: \code{\link{eq_cond_expected_outcomes}},
  \code{\link{eq_expected_outcomes}}, \code{\link{eq_li}},
  \code{\link{eq_outcomes}}, \code{\link{eq_tables}},
  \code{\link{game_gambit_solve_qre}},
  \code{\link{game_gambit_solve}},
  \code{\link{game_solve_spe}}
}
