% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{eq_tables}
\alias{eq_tables}
\title{Return solved equilibrium in a table format}
\usage{
eq_tables(game, reduce.tables = TRUE, combine = 2,
  eq.ind = seq_along(game$eq.li), ignore.keys = NULL, ...)
}
\arguments{
\item{reduce.tables}{(default = TRUE). Shall we try to reduce the rows and columns of the key tables be reduced to get a subset of neccessary keys that perfectly predict the chosen value of an action?}

\item{combine}{if 0 generate separate tables for each equilibrium. If 1 bind the tables of each variable over all equilibria. If 2 (default) also collapse the rows that are the same for different equilibria and add a column eq.inds that contains all equilibrium numbers as a comma separated string}

\item{eq.ind}{Vector of integers specifying the indices of all equilibria that shall be considered. By default all equilibria.}

\item{ignore.keys}{A character vector of variables that will always be removed from the key variables, without any check whether they are neccessary or not.}
}
\description{
Best take a look at the Vignettes to understand this format.
}
