% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambit.r
\name{expected.eq.outcomes}
\alias{expected.eq.outcomes}
\title{compute expected equilibrium outcomes
taking expectations over moves of nature}
\usage{
expected.eq.outcomes(eqo.df = NULL, group.vars = c("eq.ind", "eqo.ind"),
  eq.li = NULL, tg = NULL)
}
\description{
compute expected equilibrium outcomes
taking expectations over moves of nature
}
