% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{fill.vec}
\alias{fill.vec}
\title{fill a vector up to a specified length with fill}
\usage{
\method{fill}{vec}(vec, len = length(vec), fill = NA)
}
\description{
fill a vector up to a specified length with fill
}
