% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{flatten.names.list}
\alias{flatten.names.list}
\title{Names lists are used to recursively store order of columns}
\usage{
flatten.names.list(li, name = "")
}
\description{
Names lists are used to recursively store order of columns
}
