% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambit.r
\name{gambit.solve.eq}
\alias{gambit.solve.eq}
\title{Finds one or all mixed strategy equilibria}
\usage{
gambit.solve.eq(tg, mixed = FALSE, just.spe = TRUE,
  efg.file = tg.efg.file.name(tg), efg.dir = get.efg.dir(tg$gameId),
  gambit.dir = "", solver = NULL, eq.dir = get.eq.dir(tg$gameId),
  save.eq = TRUE, solvemode = NULL)
}
\description{
Finds one or all mixed strategy equilibria
}
