% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{game.eq.li}
\alias{game.eq.li}
\title{Return the computed equilibria using the internal representation}
\usage{
game.eq.li(game, ...)
}
\description{
Return the computed equilibria using the internal representation
}
