% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{game.eq.outcomes}
\alias{game.eq.outcomes}
\title{Return a data frame of all equilibrium outcomes}
\usage{
game.eq.outcomes(game, ...)
}
\arguments{
\item{game}{the game object for which previously equilibria were computed e.g. with \code{game_solve}.}
}
\description{
Return a data frame of all equilibrium outcomes
}
