% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{game.expected.eq.outcomes}
\alias{game.expected.eq.outcomes}
\title{Return a data frame of expected equilibrium outcomes}
\usage{
game.expected.eq.outcomes(game, ...)
}
\arguments{
\item{game}{the game object for which prevoiously equilibria were computed e.g. with \code{game_solve}.}
}
\description{
Return a data frame of expected equilibrium outcomes
}
