% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_param.r
\name{game_change_param}
\alias{game_change_param}
\title{Changes one or several parameters of a game}
\usage{
game_change_param(game, ..., params = list(),
  verbose = isTRUE(game$options$verbose > 0))
}
\description{
For an already compiled game, we try to change parameters
in a fashion that is faster than a complete recompilation.
}
\seealso{
Other Modify Game: \code{\link{game_copy}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}}

Other Game Parameters: \code{\link{make_game_params}}
}
