% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{game_copy}
\alias{game_copy}
\title{Make a deep copy of a game}
\usage{
game_copy(game)
}
\description{
Make a deep copy of a game
}
\seealso{
Other Modify Game: \code{\link{game_change_param}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}}
}
