% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{game_fix_actions}
\alias{game_fix_actions}
\title{Fix move probabilities of actions}
\usage{
game_fix_actions(game, ..., actions = list(...), tremble.prob = NULL)
}
\arguments{
\item{...}{directly the named arguments from which \code{actions} will be constructed}

\item{actions}{a named list. The names correspond to action names. The default value to fix mixed strategies is a table that specifies conditional move probabilities (see example). If you want to fix pure actions you can also provide arguments as in \code{\link{game_fix_action_preferences}}.}

\item{tremble.prob}{If a positive number, we assume that with this probability the player trembles and then chooses a random action with uniform probability. Trembles can be useful to enforce some sequential rationality in continuation play, but note that uniform trembles are not neccessarily the correct form of trembles to find sequential equilibria or trembling hand perfect equilibria.}
}
\description{
The function corresponds the provided actions into moves of nature with specified move probabilities. Can be a useful step when checking for existence of equilibria with particular structure.
}
\details{
For fixing pure strategies \code{\link{game_fix_action_preferences}} is preferable when using the \code{gambit-logit} solver that can find sequential equilibria, by using logit trembles.
}
\seealso{
Other Fix Actions: \code{\link{game_fix_action_preferences}},
  \code{\link{game_prefer_outcomes}}
}
