% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{game_prefer_outcomes}
\alias{game_prefer_outcomes}
\title{Set add a large amount of utility if a player plays a particular action}
\usage{
game_prefer_outcomes(game, player1 = NULL, player2 = NULL, player3 = NULL,
  ..., player.prefs = list(player1 = player1, player2 = player2, player3 =
  player3, ...))
}
\arguments{
\item{player1}{A formula describing which utility levels should be added to the current utility function of player 1 (see example). If NULL we don't add utilities for player 1. Similar for the other players 2-4.}

\item{...}{additional formulas for games with more than 4 players.}

\item{player.prefs}{by default equal to \code{list(player1,player2,player3, player4,...)}. Can be manually provided.}
}
\description{
Allows to study the game under the assumption that a player strongly prefers
to chose one particular move of an action variable.
}
\details{
If you want to fix mixed strategies, use the he related function \code{\link{game_fix_action_preferences}} transforms the corresponding action into a move of nature.

For fixing pure strategies \code{\link{game_fix_action_preferences}}
is preferable when using the \code{gambit-logit} solver that can
find sequential equilibria, by using logit trembles.
}
\seealso{
Other Fix Actions: \code{\link{game_fix_action_preferences}},
  \code{\link{game_fix_actions}}

Other Preferences: \code{\link{game_fix_action_preferences}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}},
  \code{\link{pref_custom}}, \code{\link{pref_envy}},
  \code{\link{pref_heterogeneous_players}},
  \code{\link{pref_ineqAv}}, \code{\link{pref_lossAv}},
  \code{\link{pref_payoff}}
}
