% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{game_write_efg}
\alias{game_write_efg}
\title{Write game as a Gambit efg file}
\usage{
game_write_efg(game, file.with.dir = file.path(dir, file),
  file = tg.efg.file.name(game$tg), dir = getwd(),
  verbose = !isTRUE(game$options$verbose == 0))
}
\arguments{
\item{game}{The game object}

\item{file.with.dir}{The file with full path. If NULL create a default name}

\item{file}{The file name without directory}

\item{dir}{The directory of a file}
}
\description{
Write game as a Gambit efg file
}
\seealso{
Other Gambit: \code{\link{game_gambit_solve}}
}
