% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.r
\name{get.project.dir}
\alias{get.project.dir}
\title{Get the gtree project directory}
\usage{
get.project.dir(xs = get.xs())
}
\description{
If the gtreeGUI is running use the project.dir specified
there.
Otherwise check getOptions("gtree.project.dir).
If that is null return the current working directory
}
