% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.r
\name{get.rg}
\alias{get.rg}
\title{Get a game in rg format by its gameId}
\usage{
get.rg(gameId = jg$gameId, jg.hash = get.jg.hash(jg = jg), jg = NULL,
  rg = NULL, games.dir = get.games.dir(project.dir),
  project.dir = get.project.dir(), save.new = TRUE)
}
\description{
If the json file has not changed return old .rg file
otherwise generate new rg file from json source
}
