% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.r
\name{get.tg}
\alias{get.tg}
\title{Get a game in tg format by its gameId}
\usage{
get.tg(variant = first.non.null(vg$variant, 1),
  gameId = first.non.null(vg$gameId, jg$gameId, rg$gameId),
  jg.hash = get.jg.hash(jg = jg, rg = rg, vg = vg), jg = NULL, rg = NULL,
  vg = NULL, tg = NULL, games.dir = get.games.dir(project.dir),
  project.dir = get.project.dir(), save.new = TRUE,
  branching.limit = 10000, msg.fun = NULL, never.load = FALSE,
  filename = NULL)
}
\description{
If the json file has not changed return old .tg file
otherwise generate new rg and tg files from json source
}
