% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.r
\name{get.vg}
\alias{get.vg}
\title{Get a game in vg format by its gameId}
\usage{
get.vg(variant = 1, gameId = first.non.null(jg$gameId, rg$gameId),
  jg.hash = get.jg.hash(jg = jg, rg = rg), jg = NULL, rg = NULL,
  vg = NULL, games.dir = get.games.dir(project.dir),
  project.dir = get.project.dir(), save.new = FALSE, always.new = FALSE)
}
\description{
If the json file has not use exsisting .rg file to extract
the vg format.
Otherwise first generate new rg files from json source
}
