% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{get_outcomes}
\alias{get_outcomes}
\title{Return a data frame of all possible outcomes}
\usage{
get_outcomes(game, reduce.cols = TRUE)
}
\arguments{
\item{game}{the game object defined with \code{new_game} and being compiled with \code{game_compile} or after a call of \code{game_solve}.}

\item{reduce.cols}{if TRUE remove some technical columns}
}
\description{
Return a data frame of all possible outcomes
}
