% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{is_true}
\alias{is_true}
\title{Returns logical vector replacing NA by FALSE}
\usage{
is_true(vec)
}
\arguments{
\item{vec}{A vector of logical values
(possible containing NA) or values that can be
transformed to logicals}
}
\description{
Returns logical vector replacing NA by FALSE
}
\seealso{
Other Helper Functions: \code{\link{case_distinction}}
}
