% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.r
\name{load.tg}
\alias{load.tg}
\title{Load tg without any checks}
\usage{
load.tg(variant = first.non.null(vg$variant, 1),
  gameId = first.non.null(vg$gameId, jg$gameId, rg$gameId), jg = NULL,
  rg = NULL, vg = NULL, tg = NULL, tg.id = tg$tg.id,
  games.dir = get.games.dir(project.dir), project.dir = get.project.dir(),
  filename = if (!is.null(tg.id)) paste0(tg.id, ".tg") else NULL)
}
\description{
Load tg without any checks
}
