% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.r
\name{make.game.dir}
\alias{make.game.dir}
\title{Create a directory structure for a new game}
\usage{
make.game.dir(gameId, games.dir = file.path(project.dir, "games"),
  project.dir = get.project.dir())
}
\description{
includes the subdirectories eq, gambit and pages
}
