% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{make_game_params}
\alias{make_game_params}
\title{Specify the game parameters}
\usage{
make_game_params(numPlayers = 2, ...)
}
\description{
This function is only to be used inside \code{new_game}. To change the parameters of an existing game call \code{game_change_params}.
}
