% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{natureMove}
\alias{natureMove}
\title{Specify a random move of nature in a stage}
\usage{
natureMove(name, set, probs = NULL, table = NULL, fixed = NULL,
  tremble.prob = NULL, ...)
}
\arguments{
\item{name}{The variable name of the variable}

\item{set}{The set of different values. Can be a rhs only formula.}

\item{probs}{The probability of each element in set. If NULL all moves are equally likely. Can be a rhs formula}
}
\description{
Specify a random move of nature in a stage
}
\seealso{
Other Define Stage: \code{\link{action}},
  \code{\link{stage}}
}
