% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{new_game}
\alias{new_game}
\title{Create a new gtree game}
\usage{
new_game(gameId, params = game_params(), options = make_game_options(),
  stages, variant = "", check = TRUE)
}
\description{
See the examples on gtree website.
}
