% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{paste.matrix.cols}
\alias{paste.matrix.cols}
\title{Paste together columns of a matrix or data.frame}
\usage{
paste.matrix.cols(mat, cols = 1:NCOL(mat), sep = "|", ...)
}
\description{
Paste together columns of a matrix or data.frame
}
