% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_change_params}
\alias{pref_change_params}
\title{Change the parameters of a preference object}
\usage{
pref_change_params(pref, ..., params = list(), label = NULL,
  players = 1:2, numPlayers = length(players))
}
\description{
Change the parameters of a preference object
}
