% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_custom}
\alias{pref_custom}
\title{Create a custom preference}
\usage{
pref_custom(..., params = NULL, label = "custom")
}
\arguments{
\item{...}{Unquoted that describe the utility as a function of the parameters of the game and possible preference parameters. Should be ordered by players. Names are irrelevant.}

\item{params}{An optional list of parameters that are used in the formulas above}

\item{label}{A label for the preference, should contain info about the parameters}

\item{type}{A general type label independet of the parameters}
}
\description{
Create a custom preference
}
\seealso{
Other Preferences: \code{\link{game_fix_action_preferences}},
  \code{\link{game_prefer_outcomes}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}},
  \code{\link{pref_envy}},
  \code{\link{pref_heterogeneous_players}},
  \code{\link{pref_ineqAv}}, \code{\link{pref_lossAv}},
  \code{\link{pref_payoff}}
}
