% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_envy}
\alias{pref_envy}
\title{Fehr-Schmidt inequality aversion with envy only}
\usage{
pref_envy(alpha = 0.75, player = 1:numPlayers, numPlayers = 2, ...)
}
\arguments{
\item{alpha}{the degree of envy}

\item{player}{player(s) for which the preferences apply. Per default 1:2}

\item{numPlayers}{number of players in game per default 2}
}
\description{
Fehr-Schmidt inequality aversion with envy only
}
