% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_heterogeneous_players}
\alias{pref_heterogeneous_players}
\title{Combine preferences for different players}
\usage{
pref_heterogeneous_players(..., prefs = list(...), label = NULL)
}
\arguments{
\item{...}{all preferences ordered by players}

\item{prefs}{alternatively the preferences as a list object}

\item{label}{optional label of preferences. If NULL the individual labels will be pasted together}

\item{type}{label of the combined preference type}
}
\description{
Combine preferences for different players
}
\seealso{
Other Preferences: \code{\link{game_fix_action_preferences}},
  \code{\link{game_prefer_outcomes}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}},
  \code{\link{pref_custom}}, \code{\link{pref_envy}},
  \code{\link{pref_ineqAv}}, \code{\link{pref_lossAv}},
  \code{\link{pref_payoff}}
}
