% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_lossAv}
\alias{pref_lossAv}
\title{'Linear loss aversion preferences with a single reference point}
\usage{
pref_lossAv(lambda = 2, r = 0, player = 1:numPlayers, numPlayers = 2)
}
\arguments{
\item{lambda}{factor by which losses loom larger than gains (default = 2)}

\item{r}{The reference point, by default 0. Can be a vector in order to have different reference points for different players.}

\item{player}{player(s) for which the preferences apply. Per default 1:2}

\item{numPlayers}{number of players in game per default 2}
}
\description{
'Linear loss aversion preferences with a single reference point
}
\seealso{
Other Preferences: \code{\link{game_fix_action_preferences}},
  \code{\link{game_prefer_outcomes}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}},
  \code{\link{pref_custom}}, \code{\link{pref_envy}},
  \code{\link{pref_heterogeneous_players}},
  \code{\link{pref_ineqAv}}, \code{\link{pref_payoff}}
}
