% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_payoff}
\alias{pref_payoff}
\title{Utility is equal to monetary payoff.}
\usage{
pref_payoff(player = 1:2, ...)
}
\arguments{
\item{player}{player(s) for which the preferences apply. Per default 1:2}
}
\description{
This means the player is simply a risk
neutral expected payoff maximizer.
}
