% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r, R/tools.r
\name{sc}
\alias{sc}
\alias{sc}
\title{Like paste0 but returns an empty vector if some string is empty}
\usage{
sc(..., sep = "", collapse = NULL)

sc(..., sep = "", collapse = NULL)
}
\description{
Like paste0 but returns an empty vector if some string is empty

Like paste0 but returns an empty vector if some string is empty
}
