% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_spe.R
\name{solve.sg.spe}
\alias{solve.sg.spe}
\title{solve all spe of subgame .sg.ind
assumes that descendent subgames have already been solved
and uses backward induction}
\usage{
\method{solve}{sg.spe}(.sg.ind = 1, tg)
}
\description{
solve all spe of subgame .sg.ind
assumes that descendent subgames have already been solved
and uses backward induction
}
