% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{stage}
\alias{stage}
\title{Specify a stage for a game}
\usage{
stage(name, player = NULL, condition = NULL, observe = NULL,
  compute = NULL, nature = NULL, actions = NULL, ...)
}
\arguments{
\item{name}{Name of the stage}

\item{player}{The player who acts in this stage. Can be a rhs formula. If an action is chosen in the stage, there must be a unique player. If it is a stage in which no actions take place, the player variable multiple players can be set. Each player observes the variables specified under observe.}

\item{condition}{A logical condition specifying whether the stage will be run. Can be a rhs formula. If it evaluates to FALSE the stage will not be shown, i.e. no observations are made and no actions are chosen. Also no computations in this stage take place.}

\item{observe}{A vector of variable names specifying which variables are observed by the player(s) at this stage. Is relevant to correctly specify the information sets in the extensive form game.}

\item{compute}{A list of formulas like 'compute=list(payoff_1 ~ x-5)'. The lhs specifies a variable name and the rhs a DETERMINISTIC formula. The variables are computed at the beginning of the stage before actions and moves of nature take place. This means they can be used e.g. in formulas for action sets of the same stage.}

\item{nature}{A list of moves of nature, i.e. random variables from a finite set. E.g. nature=list(natureMove("proposer",c(1,2),prob=c(0.4,0.6)).}

\item{actions}{A list of actions. E.g. actions=list(action("offer",~0:cake_size)}
}
\description{
Specify a stage for a game
}
