% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{str.combine}
\alias{str.combine}
\title{Like paste0 but returns an empty vector if some string is empty}
\usage{
\method{str}{combine}(..., sep = "", collapse = NULL)
}
\description{
Like paste0 but returns an empty vector if some string is empty
}
