% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{str.space}
\alias{str.space}
\title{Returns a string constisting of times spaces, vectorized over times}
\usage{
\method{str}{space}(times, space = " ")
}
\description{
Returns a string constisting of times spaces, vectorized over times
}
