% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{var.in.expr}
\alias{var.in.expr}
\title{Returns all variable names in an R expression}
\usage{
var.in.expr(expr, expr.str, envir = baseenv())
}
\description{
Returns all variable names in an R expression
}
