% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{var.in.expr.str}
\alias{var.in.expr.str}
\title{Returns all variable names in an R expression}
\usage{
var.in.expr.str(expr.str, envir = baseenv(), union = TRUE)
}
\description{
Returns all variable names in an R expression
}
