% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{variants.from.tree.path}
\alias{variants.from.tree.path}
\title{Gets game variants that correspond to a tree path}
\usage{
variants.from.tree.path(tree.path)
}
\description{
Gets game variants that correspond to a tree path
}
