% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eedb_utils.r
\name{vec.list.to.matrix}
\alias{vec.list.to.matrix}
\title{Comverts a list of vectors into a matrix, shorter vectors will be filled up}
\usage{
vec.list.to.matrix(li, fill = NA, transpose = TRUE)
}
\description{
Comverts a list of vectors into a matrix, shorter vectors will be filled up
}
