% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{eq_expected_outcomes}
\alias{eq_expected_outcomes}
\title{Return a data frame of expected equilibrium outcomes}
\usage{
eq_expected_outcomes(game, ...)
}
\arguments{
\item{game}{the game object for which prevoiously equilibria were computed e.g. with \code{game_solve}.}
}
\description{
Return a data frame of expected equilibrium outcomes
}
