% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_param.r
\name{game_change_param}
\alias{game_change_param}
\title{Changes one or several parameters of a game}
\usage{
game_change_param(game, ..., params = list(),
  verbose = isTRUE(game$options$verbose > 0))
}
\description{
Changes one or several parameters of a game
}
