% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{game_gambit_solve_qre}
\alias{game_gambit_solve_qre}
\title{Solve for quantal response equilibria using Gambit}
\usage{
game_gambit_solve_qre(game, gambit.command = "gambit-logit -q -l",
  gambit.dir = "", efg.file = NULL, efg.dir = NULL,
  verbose = isTRUE(game$options$verbose >= 1))
}
\description{
This function computes logit agent quantal response equilibria using the Gambit solver gambit-logit. For a short description see the \href{https://en.wikipedia.org/wiki/Quantal_response_equilibrium}{Wikipedia article} and the gambit-logit solver's \href{documentation}{https://gambitproject.readthedocs.io/en/latest/tools.html#gambit-logit-compute-quantal-response-equilbria}. Details are in the article \href{Using Quantal Response to Compute Nash and Sequential Equilibria}{https://link.springer.com/article/10.1007/s00199-009-0443-3} by Theodore Turocy. But unfortunately, the article can only be found behind a pay wall.
}
\details{
For a description of the arguments see \code{\link{game_gambit_solve}}
}
\seealso{
Other eq: \code{\link{eq_cond_expected_outcomes}},
  \code{\link{eq_cond_outcomes}},
  \code{\link{eq_expected_outcomes}}, \code{\link{eq_li}},
  \code{\link{eq_outcomes}}, \code{\link{eq_tables}},
  \code{\link{game_gambit_solve}},
  \code{\link{game_solve_spe}}
}
