% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_change_params}
\alias{pref_change_params}
\title{Change the parameters of a preference object}
\usage{
pref_change_params(pref, ..., params = list(), label = NULL,
  players = 1:2, numPlayers = length(players))
}
\description{
Change the parameters of a preference object
}
\seealso{
Other Modify Game: \code{\link{game_change_param}},
  \code{\link{game_copy}},
  \code{\link{game_set_preferences}}

Other Preferences: \code{\link{game_fix_action_preferences}},
  \code{\link{game_prefer_outcomes}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_custom}}, \code{\link{pref_envy}},
  \code{\link{pref_heterogeneous_players}},
  \code{\link{pref_ineqAv}}, \code{\link{pref_lossAv}},
  \code{\link{pref_payoff}}
}
