% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_heterogeneous_players}
\alias{pref_heterogeneous_players}
\title{Combine preferences for different players}
\usage{
pref_heterogeneous_players(..., prefs = list(...), label = NULL)
}
\arguments{
\item{...}{all preferences ordered by players}

\item{prefs}{alternatively the preferences as a list object}

\item{label}{optional label of preferences. If NULL the individual labels will be pasted together}

\item{type}{label of the combined preference type}
}
\description{
Combine preferences for different players
}
