% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_ineqAv}
\alias{pref_ineqAv}
\title{Fehr-Schmidt inequality aversion.}
\usage{
pref_ineqAv(alpha = 0.75, beta = 0.5, player = 1:numPlayers,
  numPlayers = 2, ...)
}
\arguments{
\item{alpha}{the degree of envy}

\item{beta}{the degree of guilt}

\item{player}{player(s) for which the preferences apply. Per default 1:2}

\item{numPlayers}{number of players in game per default 2}
}
\description{
Fehr-Schmidt inequality aversion.
}
\seealso{
Other Preferences: \code{\link{game_fix_action_preferences}},
  \code{\link{game_prefer_outcomes}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}},
  \code{\link{pref_custom}}, \code{\link{pref_envy}},
  \code{\link{pref_heterogeneous_players}},
  \code{\link{pref_lossAv}}, \code{\link{pref_payoff}}
}
