% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\name{pref_payoff}
\alias{pref_payoff}
\title{Utility is equal to monetary payoff.}
\usage{
pref_payoff(player = 1:2, ...)
}
\arguments{
\item{player}{player(s) for which the preferences apply. Per default 1:2}
}
\description{
This means the player is simply a risk
neutral expected payoff maximizer.
}
\seealso{
Other Preferences: \code{\link{game_fix_action_preferences}},
  \code{\link{game_prefer_outcomes}},
  \code{\link{game_set_preferences}},
  \code{\link{pref_change_params}},
  \code{\link{pref_custom}}, \code{\link{pref_envy}},
  \code{\link{pref_heterogeneous_players}},
  \code{\link{pref_ineqAv}}, \code{\link{pref_lossAv}}
}
